Introduced alphabetic sorting
if author's surname is given it will be the first element to be sorted,
otherwise author's name will be taken. So if both surname's are give
they will be matched against each other otherwise it may be possibile
that a surname will be matched against a name. System tries his to
provide the best sort